adds in a11y text for state of each SR choice; adds disabled to the SR choices; moves aria-hidden higher to hide buttons; removes a11y text from several choice icon files
