+++ packages/perseus/src/widgets/radio/choice-icon/choice-icon.jsx
@@ -2,7 +2,6 @@
 /* Component for rendering a letter icon in radio choice */
 
 import Color from "@khanacademy/wonder-blocks-color";
-import * as i18n from "@khanacademy/wonder-blocks-i18n";
 import * as React from "react";
 
 import {getChoiceLetter} from "../util.js";
@@ -34,64 +33,6 @@ type ChoiceIconProps = {|
     primaryProductColor?: string,
 |};
 
-function getA11yText(
-    letter: string,
-    checked: boolean,
-    correct: boolean,
-    crossedOut: boolean,
-    showCorrectness: boolean,
-): string {
-    // There are two pieces of metadata we want to add to each a11yText:
-    // whether the answer was checked/crossed-out/neither, and whether the
-    // answer is correct/incorrect/not-yet-revealed.
-    //
-    // Translation is tricky for cross-product situations like this, so
-    // we've just enumerated all 9 possibilities as separate strings.
-    if (showCorrectness && correct) {
-        if (checked) {
-            return i18n._("(Choice %(letter)s, Checked, Correct)", {
-                letter,
-            });
-        }
-        if (crossedOut) {
-            return i18n._("(Choice %(letter)s, Crossed out, Correct)", {
-                letter,
-            });
-        }
-        return i18n._("(Choice %(letter)s, Correct)", {
-            letter,
-        });
-    }
-    if (showCorrectness && !correct) {
-        if (checked) {
-            return i18n._("(Choice %(letter)s, Checked, Incorrect)", {
-                letter,
-            });
-        }
-        if (crossedOut) {
-            return i18n._("(Choice %(letter)s, Crossed out, Incorrect)", {
-                letter,
-            });
-        }
-        return i18n._("(Choice %(letter)s, Incorrect)", {
-            letter,
-        });
-    }
-    if (checked) {
-        return i18n._("(Choice %(letter)s, Checked)", {
-            letter,
-        });
-    }
-    if (crossedOut) {
-        return i18n._("(Choice %(letter)s, Crossed out)", {
-            letter,
-        });
-    }
-    return i18n._("(Choice %(letter)s)", {
-        letter,
-    });
-}
-
 function ChoiceIcon(props: ChoiceIconProps): React.Node {
     const {
         pos,
@@ -111,19 +52,11 @@ function ChoiceIcon(props: ChoiceIconProps): React.Node {
     } = props;
 
     const letter = getChoiceLetter(pos);
-    const a11yText = getA11yText(
-        letter,
-        checked,
-        correct,
-        crossedOut,
-        showCorrectness,
-    );
 
     if (product === "sat") {
         return (
             <SATChoiceIcon
                 letter={letter}
-                a11yText={a11yText}
                 reviewMode={reviewMode}
                 checked={checked}
                 correct={correct}
@@ -135,7 +68,6 @@ function ChoiceIcon(props: ChoiceIconProps): React.Node {
     return (
         <LibraryChoiceIcon
             letter={letter}
-            a11yText={a11yText}
             reviewMode={reviewMode}
             checked={checked}
             crossedOut={crossedOut}
